About the work of code
This code is written to filter and copy the data from Sheet1 on the basis of each different type of value in column "B" then create a new excel file paste the data in it and remove all column having sum = 0 and apply sum function on last column and just adjacent cell of sum value write "Total". Finally save the file on name of value of column "B" the unique on the basis the filter was applied.